Refactoring the DoNotEncoder out to avoid duplication
