+++ lib/facilities/ppms/do_not_encoder.rb
@@ -1,0 +1,16 @@
+module Facilities
+  module PPMS
+
+    # ppms has strange behavior for certain url-encoded characters, no url-encoding works best
+    class DoNotEncoder
+      def self.encode(params)
+        buffer = +''
+        params.each do |key, value|
+          buffer << "#{key}=#{value}&"
+        end
+        buffer.chop
+      end
+    end
+
+  end
+end
