+++ lib/facilities/ppms/v0/configuration.rb
@@ -3,6 +3,7 @@
 require 'common/client/configuration/rest'
 require 'common/client/middleware/response/raise_error'
 require 'common/client/middleware/response/ppms_parser'
+require 'facilities/ppms/do_not_encoder'
 
 module Facilities
   module PPMS
@@ -17,22 +18,12 @@ module Facilities
         def service_name
           'PPMS'
         end
-        # ppms has strange behavior for certain url-encoded characters, no url-encoding works best
-        class DoNotEncoder
-          def self.encode(params)
-            buffer = +''
-            params.each do |key, value|
-              buffer << "#{key}=#{value}&"
-            end
-            buffer.chop
-          end
-        end
 
         def connection
           Faraday.new(base_path, headers: base_request_headers, request: request_options) do |conn|
             conn.use :breakers
             conn.use :instrumentation, name: 'facilities.ppms.request.faraday'
-            conn.options.params_encoder = DoNotEncoder
+            conn.options.params_encoder = Facilities::PPMS::DoNotEncoder
 
             # Uncomment this if you want curl command equivalent or response output to log
             # conn.request(:curl, ::Logger.new(STDOUT), :warn) unless Rails.env.production?
